Additional test for demonstrating possible issues
This test was suggested to me personally in order to demonstrate a basic real
life difficulty when handling dates around specific "magic" dates - in this case
the original transition period from Julian to Gregorian calendar in October 1582
which literally is supposed to "skip" 10 days.

JavaScript's native Date-object does not handle Julian dates; instead, it does
represent what the dates would have been if the Gregorian calendar had been in
use non-stop throughout history. This is generally just fine for most real life
cases where calendar only needs to show accurate dates not further back in the
past than four and a half century, which generally is "good enough" for most
developers of modern software.

For historical accuracy, though, this can definitely lead to problems. However,
I believe that in such cases the date value is most likely the least of the
issues that need to be thought out. Some additional issues I can easily come
up with off the top of my head are:

* handling chinese, japanese, buddhist, muslim, hebrew, etc. calendars (that
  are still widely used around the world)
* how to visually represent what calendar is in use
* mixing & matching the various calendars

That is a can of worms I don't really feel comfortable opening at this time.

Because of this, I created this one test to demonstrate that there are known
limitations in the package; they are known, but not addressed in implementation.